gen_wb_wrapper_hdl: fix regular reg. assignment in hierarchical newad modules
Previously, we were trying to find port name
in the to-be-wrapper module that matched with
a signal in wishbone CSR module.

That only works in the case of leaf newad modules
that the names do not have any prefixes in them,
but fails for hierarchical modules.

Instead of trying to match the names just use
the one generated by cheby. By our naming rules
it will automatically match.